refactor: split Carousel components on different files
